+++ components/menu/PadMenu.vue
@@ -2,11 +2,11 @@
 </script>
 
 <template>
-  <div class="border-gray-500 bg-gray-800 p-3 w-1/2 mx-auto mt-10 rounded-xl border-2 shadow-lg shadow-void-200">
+  <div class="mx-auto mt-10 w-1/2 rounded-xl border-2 border-gray-500 bg-gray-800 p-3 shadow-lg shadow-gray-400">
     <slot />
   </div>
-  <div class="text-center mt-4">
-    Created by <a class="text-gray-700 hover:text-gray-500" target="_blank" href="https://github.com/tomaszbawor">Tomasz Bawor</a>
+  <div class="mt-4 text-center">
+    Created by <a class="text-gray-900 hover:text-gray-500" target="_blank" href="https://github.com/tomaszbawor">Tomasz Bawor</a>
   </div>
 </template>