+++ components/ui/IconButton.vue
@@ -1,22 +1,12 @@
 <script setup lang="ts">
-import { computed } from 'vue'
-
-const props = defineProps<{
-  variant: 'primary' | 'secondary'
-  isCircle?: boolean
-}>()
-
 defineEmits(['onClick'])
-
-const buttonClass = computed(() => {
-  return props.variant === 'primary'
-    ? 'bg-gray-400 rounded-full text-white px-4 py-2 rounded px-auto'
-    : 'bg-gray-500 rounded-full text-white px-4 py-2 rounded'
-})
 </script>
 
 <template>
-  <button :class="buttonClass" @click="$emit('onClick')">
+  <button
+    type="button" class="mr-2 inline-flex items-center rounded-full bg-gray-500 p-2.5 text-center text-sm font-medium text-white hover:bg-blue-800 focus:outline-none focus:ring-4 focus:ring-blue-300 dark:bg-gray-600 dark:hover:bg-gray-700 dark:focus:ring-blue-800"
+    @click="$emit('onClick')"
+  >
     <slot />
   </button>
 </template>