+++ layouts/game.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+</script>
+
+<template>
+  <GameContainer>
+    <div class="flex flex-row">
+      <GameSidebar />
+      <div>
+        <slot />
+      </div>
+    </div>
+  </GameContainer>
+</template>