+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.2](https://github.com/tomaszbawor/padron/compare/v0.1.1...v0.1.2) (2023-09-17)
+
+
+### Features
+
+* add calculation for exp needed for next level ([0a253fa](https://github.com/tomaszbawor/padron/commit/0a253fa1079a256abb6afb646896e013b82a0f81))
+* add flowbite components ([e7ed9c0](https://github.com/tomaszbawor/padron/commit/e7ed9c0947bbdb25e71bd107d76c352c1fb3a8f8))
+* add hero card on side menu ([065b2e8](https://github.com/tomaszbawor/padron/commit/065b2e8604d803e4a97c56acbd4f1250cdb6c2c7))
+* add navbar ([bd89e02](https://github.com/tomaszbawor/padron/commit/bd89e02f3eaf27ad3572e9ab1e5e12c904f93c8b))
+* add posibility to specify hero name ([403fa99](https://github.com/tomaszbawor/padron/commit/403fa99257f80865e3c8aa128b01808a71b56259))
+* add tailwind eslint config ([15e3b85](https://github.com/tomaszbawor/padron/commit/15e3b85a9b46ffe7e7ead24a55e024aea6587462))
+* add vercel analytics ([c472a19](https://github.com/tomaszbawor/padron/commit/c472a19c848cc17044b047f5dbda53f367a6b1bb))
+* adding experience and level to hero ([32d1eac](https://github.com/tomaszbawor/padron/commit/32d1eac258eb20396f983a4333abfe85323049a6))
+* continue disabled when no save provided, and add label with hero name and level ([17cac1e](https://github.com/tomaszbawor/padron/commit/17cac1e372a94ca4dba9a6927bfbb3c9212e4f2f))
+* make Mana and Health constant at start, no skillpoint needed ([17fa5b2](https://github.com/tomaszbawor/padron/commit/17fa5b291744e706c361a6c58662629c806778ea))
+
 ### [0.1.1](https://github.com/tomaszbawor/padron/compare/v0.1.0...v0.1.1) (2023-09-15)